I guess this can be a float even if it's 0 to 100 (0.0 to 100.0)

+++ src/lib/datadog/datadog-apm.ts
@@ -1,7 +1,7 @@
 import tracer from 'dd-trace'
 
 function parseSampleRate(rateString: string | undefined): number | undefined {
-  const rate = parseInt(rateString)
+  const rate = parseFloat(rateString)
   if (isNaN(rate)) {
     return undefined
   }
